Add chat context collector
