+++ Core/Sources/ChatContextCollector/ChatContextCollector.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol ChatContextCollector {
+    func generateSystemPrompt(oldMessages: [String]) -> String
+}
